Add Cancel Purchase Command
