+++ Shoping.Application/Features/Purchases/Queries/GetPurchasesQuery.cs
@@ -25,7 +25,9 @@ public class GetPurchasesQueryHandler : IRequestHandler<GetPurchasesQuery, List<
 
     public async Task<List<GetPurchasesQueryResponse>> Handle(GetPurchasesQuery request, CancellationToken cancellationToken)
     {
-        var purchases = await _unitOfWork.Purchase.GetAllAsync(p => p.IsDeleted == false);
+        var purchases = await _unitOfWork.Purchase.GetAllAsync(p => p.IsCanceled == false);
+        // var purchases = await _unitOfWork.Purchase.GetAllAsync(p => p.IsDeleted == false);
+
         return _mapper.Map<List<GetPurchasesQueryResponse>>(purchases);
     }
